SREX-1440 distinguish environments in generated sentry issues
for now, we'll make the assumptions that we are segreggating by the environment.

Sentry's api does not support fetching the tag details of the issues, so we cannot
dynamically generate the environment label from the response.

We'd have to iterate over each issue and call the fetch tag api in order to
achieve that. However, it'll be a serious performance hit if we did that, and
it's not really feasible with the volume of issues we have.